Fix possible use-after-free when using stream* and raw* encoders
Previously, it was possible that during reallocation of FiltersArray::options,
undefined behaviour resulted because the pointers in the filter structures
were not updated and used later on.
Using std::list instead of std::vector works just as well and ensures
referential integrity.
This bug only shows up when using the `.filters` option with more than
one filter.