Reset uv_async structure when coding is done, count processed chunks
Resetting the outputDataAsync is to make sure that node.js
does not believe that any data streams are still working
or will push async evvents.

Counting of processed chunks is to provide a better overview
of which chunks have been processed and invoking the callbacks
provided to _transform accordingly.

Conflicts:
	src/lzma-stream.cpp